ajustes

+++ client-proj/src/components/ms-person.vue
@@ -1,6 +1,5 @@
 
 <template>
-
     <div class="q-pa-md row items-start q-gutter-md">
       <q-card class="my-card">
         <q-card-section>
@@ -10,7 +9,7 @@
                 <div><q-input outlined v-model="data" :dense="dense" placeholder="Nome" hint="Data de nascimento" mask="##/##/####" fill-mask class="nasc"/></div>
                 <div><q-input outlined v-model="cpf" :dense="dense" placeholder="C.P.F" hint="C.P.F" mask="###.###.###-##" fill-mask class="cpf"/></div>
                 <div class="q-gutter-md">
-                    <q-select outlined v-model="nacionalid" :dense="dense" :options="options" label="Nacionalidade" hint="Nacionalidade" style="width: 189px"/>
+                    <q-select outlined v-model="nacionalid" option-label="nome" option-value="id" :dense="dense" :options="options" label="Nacionalidade" hint="Nacionalidade" style="width: 189px"/>
                 </div>
                 
                 <br /><br /><br />
@@ -52,13 +51,44 @@
             <q-card-section v-if="rows.length > 0">
                 <div class="row2">
                     &nbsp;&nbsp;&nbsp;
-                    <q-btn color="white" text-color="black" label="Excluir" icon="delete" :disable="selected.length == 0" @click="deletePerson()"/>&nbsp;&nbsp;&nbsp;
-                    <q-btn color="white" text-color="black" label="Alterar" icon="edit" :disable="selected.length > 1 || selected.length == 0"/>
+                    <q-btn color="white" text-color="black" label="Excluir" icon="delete" :disable="selected.length == 0" @click="deletePerson(true)"/>&nbsp;&nbsp;&nbsp;
+                    <q-btn color="white" text-color="black" label="Alterar" icon="edit" :disable="selected.length > 1 || selected.length == 0" @blur="alterarPerson=false" @click="alterar()"/>
                 </div>
             </q-card-section>
         </div>
 
       </q-card>
+
+      <q-dialog v-model="confirm" persistent>
+      <q-card>
+        <q-card-section class="row items-center">          
+          <span class="q-ml-sm">Tem certeza que deseja exluir o(s) registro(s)?</span>
+        </q-card-section>
+
+        <q-card-actions align="right">
+          <q-btn flat label="Sim" color="primary" v-close-popup @click="deletePerson(false)"/>
+          <q-btn flat label="Cancelar" color="primary" v-close-popup @click="excluirOk=false"/>
+        </q-card-actions>
+      </q-card>
+    </q-dialog>
+
+    <q-dialog v-model="alert">
+      <q-card>
+        <q-card-section>
+          <div class="text-h6">Erro</div>
+        </q-card-section>
+
+        <q-card-section class="q-pt-none">
+          {{ mensagemErro }}
+        </q-card-section>
+
+        <q-card-actions align="right">
+          <q-btn flat label="OK" color="primary" v-close-popup />
+        </q-card-actions>
+      </q-card>
+    </q-dialog>
+
+    <ms-person-alter v-if="alterarPerson" :nomePerson="objetoPerson" v-on:close-form="getList()"/>
 
     </div>
 </template>
@@ -67,6 +97,7 @@
 import { ref } from 'vue'
 import axios from 'axios'
 import { Buffer } from 'buffer';
+import msPersonAlter from './ms-person-alter.vue';
 
 const columns = [
   { id: 'id', label: 'Id', field: 'id', sortable: true, align: 'left' },
@@ -83,7 +114,7 @@ const columns = [
   { name: 'cpf', label: 'C.P.F', field: 'cpf' , align: 'left'},
   { name: 'nomeMae', label: 'Nome da mãe', field: 'nomeMae', align: 'left' },
   { name: 'nomePai', label: 'Nome do pai', field: 'nomePai' , align: 'left'},
-  { name: 'notionality', label: 'Nacionalidade', field: 'notionality', sortable: true , align: 'left'},
+  { name: 'nationality', label: 'Nacionalidade', field: 'nationality', sortable: true , align: 'left'},
   
 ]
 
@@ -93,6 +124,7 @@ const password = 'e1e96229eddcbe05f18939aa47c0fdcbec08863e';
 const authHeader = 'Basic ' + Buffer.from(username + ':' + password).toString('base64');
 
 export default {
+  components: { msPersonAlter },
     
 
   data:function(){
@@ -107,8 +139,13 @@ export default {
         nomePai: ref(''),
         nacionalid: ref(''),
         dense: ref(true),
-        options: ['Brasileiro'],
-        myColums :["nome","nascimento","cpf","nomeMae","nomePai"]
+        options: [{ id:1,nome:'Brasileiro'},{ id:2,nome:'Argentino'}],
+        myColums :["nome","nascimento","cpf","nomeMae","nomePai"],
+        excluirOk: ref(false),
+        mensagemErro: ref(''),
+        alert: ref(false),
+        alterarPerson: ref(false),
+        objetoPerson: null
        
     }
   },
@@ -117,7 +154,7 @@ export default {
     
     return {
       
-      
+        confirm: ref(false),
     }
   },
   methods: {
@@ -136,6 +173,10 @@ export default {
         }).then((response) => {
             
             for (let index = 0; index < response.data.content.length; index++) {
+
+                const dataOriginal = response.data.content[index].nascimento;
+                const dataConvertida = dataOriginal.split("-").reverse().join("/");
+                response.data.content[index].nascimento = dataConvertida;
                 this.rows.push(response.data.content[index]);
                 
             }
@@ -157,7 +198,7 @@ export default {
             nomePai: this.nomePai,
             cpf: cpfNumerico,
             nationality:{
-                id:1
+                id: this.nacionalid.id
             }
         }, {
             headers: {
@@ -170,32 +211,53 @@ export default {
         })
         .catch(error => {
             console.log(error);
+            this.mensagemErro = error.response.data.detail;
+            this.alert = true;
         });       
     },
 
-    async deletePerson() {
-        
-    /* eslint-disable */
+    async deletePerson(varConfirm) {
 
-        for (let index = 0; index < selected.value.length; index++) {  
-
-           await axios.delete(`http://localhost:8080/v1/persons/${selected.value[index].id}`, {
-            headers: {
-                Authorization: authHeader
-            }
-            })
-            .then(response => {
-                console.log(response.data);                                       
-            })
-            .catch(error => {
-                console.log(error);
-            });
-            
+        if(varConfirm){
+            this.confirm = true;
+        }else{
+            this.excluirOk = true;
         }
 
-        this.getList();
+    /* eslint-disable */
+
+        if (this.excluirOk){
+
+            this.alert = false
+
+            for (let index = 0; index < selected.value.length; index++) {  
+
+            await axios.delete(`http://localhost:8080/v1/persons/${selected.value[index].id}`, {
+                headers: {
+                    Authorization: authHeader
+                }
+                })
+                .then(response => {
+                    console.log(response.data);    
+                    this.excluirOk =false;  
+                    this.confirm = false;                                 
+                    this.getList();
+                    this.selected = []
+                })
+                .catch(error => {
+                    console.log(error);
+                    this.excluirOk =false;
+                    this.confirm =false;
+                });
+                
+            }
+        }
         
-    },    
+    } ,
+    alterar(){
+        this.objetoPerson =this.selected[0];
+        this.alterarPerson=true;
+    }  
   },
   mounted(){
